[MAG-85] Adjusting Configuration Paths
Re-adjusting configuration paths to make those working with scopes.
Without those adjustments weird full paths (which are not in the database)
are formed during read/retrieval.
The adjustments are necessary (even though there is no default) so that
after a user modifies configuration values those values will be saved in
 `core_config_data` table with respect to those paths. These values can be
 retrieved later without problems.