MAG-100: Handling precision of currencies
For currencies other than the one for Japan we are using two decimal places.
Otherwise a buyer's attempt to pay would fail.
Had to put Ryan's MAG-75 code back to be able to test in development setup.
There are lot of duplicate code in PayPal and Credit card related processing.
Refactoring those should be dealt with as a seperate ticket.