Magento 102 (#55)
* MAG-100: Handling precision of currencies

For currencies other than the one for Japan we are using two decimal places.
Otherwise a buyer's attempt to pay would fail.
Had to put Ryan's MAG-75 code back to be able to test in development setup.
There are lot of duplicate code in PayPal and Credit card related processing.
Refactoring those should be dealt with as a seperate ticket.

* [MAG-100]: Adjustment in Currency in PayPal code

There are two model files related to Paypal.
One was not changed and commited previously.
Committing it now.

* [MAG-101] Making sure that a configurable product is not counted twice

Magento (by design) inserts two rows in the database per configurable product.
Our existing code was counting such a product like pant twice even if the buyer
had orderd one pant. This bug could not be located in the past due to another
bug coming in the middle/obscuring it (MAG-100).
This ticket fixes the `double counting` issue.

* [MAG-102] Fix to properly update the `reach_currency_rate` table

`Fix for first half of the issue`.
Due to a bug in the code of Magento persistence layer most rows in
our `reach_currency_rate` table were never getting updated beyond the
first time insert. This would cause payment failure during the checkout
 process (in our Magento extension). The fix in this commit is one
possible workaround.
See also: https://github.com/magento/magento2/issues/4174

* Cleaning up

Note: This code is not related to MAG-102 fix. It is just getting
 committed/included in that branch. It is needed to test
payment (in local development environment) - more specifically for
MAG-100 and MAG-101. MAG-100 and MAG-101 should go out together with MAG-102
as part of the same release.

* Cleanup as per Code Review Feedback

* Added explanation on how to deal with a configurable product

* Removing Redundant Call to a Method

* Removing Redundant Method Call