[MAG-25] Fix so that Tax in Base Currency is Saved in Database and Displayed on UI
Without this fix tax was getting saved only in the display currency not both in the
 display currency and the base currency.

With this fix tax in base currency would be saved and be visible in Sales->Order,
Sales->Invoice and Sales->Creditmemo views in the Magento Admin.

See also MAG-84 which is a duplicate of MAG-25.