Fixing refunds for PayPal
- Refunds for PayPal was not working at all, regardless of the currency
used. The underlying issue is that when a request is made to the /refund
endpoint the OrderId and ReferenceId provided are not provided as
expected. Therefore the request was failing. This fix ensures that the
correct values are stored for the original payment made using Reach
PayPal, which would allow the refund request to have the correct values
associated to the refund request.